ref: 워크스페이스  api 정리 및 import 구문 수정
