+++ src/components/common/dropdown/Filter.Style.ts
@@ -49,7 +49,7 @@ export const Title = styled.div`
 
 export const FilterSection = styled.div`
   display: grid;
-  grid-template-columns: repeat(3, 1fr); // 정확히 3열
+  grid-template-columns: repeat(3, 1fr);
   gap: 10px 24px;
   width: 100%;
 `;
